Make `make lint` pass, and gate precommit on it
Two halves.

1) Fix what should be fixed (mechanical, verified):
   - C408 x13: dict(...) -> {...} literals (robot.py BAM kwargs, blend/gate
     tables). Values unchanged; _BAM_ACTUATOR_KWARGS still expands identically.
   - SIM401: sensors[k] if k in sensors else default -> sensors.get(k, default)
   - RUF046: redundant int(round(x)) -> round(x)
   - PLW1510: explicit check=False on the nvidia-smi probe
   - EXE002: dropped a stray executable bit on src/infer.py (no shebang)

2) Ignore what is intentional, with the reason recorded in pyproject.toml
   rather than argued again later:
   - B008  mjlab's SceneEntityCfg-as-default idiom, used by every mdp function
   - BLE001 + S110  deliberate best-effort guards around the optional Patch 4
     monkey-patch and the hardware/IO paths in infer.py / sim_body_server.py.
     Narrowing these changes runtime failure modes on the robot for no gain.
   - SIM102  nested hasattr/sensor-present checks read as two preconditions
   - RUF012  class-level constant lookup tables that are never mutated

Everything not ignored is now expected to pass, so precommit runs lint again
(it was skipped while the backlog stood at 151).

On this branch alone  reports the 9 findings that the two
companion PRs remove (F811/F841/RUF059); with all three merged it is clean.
187 passed, 1 skipped.

+++ src/infer.py
@@ -1073,7 +1073,7 @@ def main():
     # to print a running average so we can compare commanded vs achieved speed.
     from collections import deque
 
-    _vel_window_steps = max(1, int(round(1.0 / control_dt)))  # ≈ 50 @ 50 Hz
+    _vel_window_steps = max(1, round(1.0 / control_dt))  # ≈ 50 @ 50 Hz
     vel_history = deque(maxlen=_vel_window_steps)
 
     csv_data = [] if args.save_csv else None
